+++ src/OpenSolution.ps1
@@ -1,4 +1,4 @@
- <#
+<#
  .Synopsis
   Opens the first visual studio solution found in the current directory.
 
@@ -12,25 +12,21 @@
    # Open the solution located in the current directory using the alias.
    sln
 #>
-function Open-Solution
-{
+function Open-Solution {
     Set-StrictMode -Version 2.0
-	$currentDirectory = Get-Location
+
+    $currentDirectory = Get-Location
   
-    $numberOfSolutionFiles = Get-ChildItem -Name | Where-Object { $_ -match '.sln$' } | Measure-Object | %{$_.Count}  
-   
+    $numberOfSolutionFiles = Get-ChildItem -Name | Where-Object { $_ -match '.sln$' } | Measure-Object | ForEach-Object { $_.Count }   
     
-    if ($numberOfSolutionFiles -eq 0) 
-    {
+    if ($numberOfSolutionFiles -eq 0) {
         Write-Host 'No solution file found'
     }
-    elseif ($numberOfSolutionFiles -eq 1) 
-    {
+    elseif ($numberOfSolutionFiles -eq 1) {
         $solutionFile = Get-ChildItem $currentDirectory -Name | Where-Object { $_ -match '.sln$' }
         Write-Host "Opening $solutionFile"
     }
-    else 
-    {
+    else {
         $solutionFile = Get-ChildItem $currentDirectory -Name | Where-Object { $_ -match '.sln$' } | Select-Object -First 1
         Write-Host "Multiple solution files found. Opening $solutionFile"
     }