+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # PSDevtools Release History
 
+## 1.0.2 - April 8, 2019
+
+### Fixed
+
+- Open-Solution now doesn't error when there is no solution file
+
+### added
+- Set-FileDates, alias touch
+
 ## 1.0.0 - April 7, 2019
 
 ### Initial Release